requirements.txt
biopython==1.86
numpy==2.3.3
pandas==2.3.2
scipy==1.16.2
PyYAML==6.0.2
requests==2.32.5
anarci==2021.02.04
freesasa==2.2.1
peptides==0.5.0
propy3==1.1.1
pandaprot==0.6.0
